mostly working w dimensional mismatch
I promise it was working and then I accidentally did something and the training data for the input lost the 'CO' column and I couldn't figure out where that happened. It should be a fairly quick fix if the words aren't dancing on my screen... I guess it's bed time!